feat : post list - 관심사 분리 리팩토리 진행
